fixing rbfDDA bug:
library(RSNNS)
set.seed(2)

nr <- 2000
nc <- 100;
inputs <- matrix(runif(nr*nc), nrow=nr, ncol=nc);
targets <- sample(x=c(1,2,3,4), size=nr, replace=TRUE);
decTargets <- decodeClassLabels(targets);

model <- rbfDDA(inputs, decTargets)

model$snnsObject$getUnitBias(201)

+++ src/kr_mem.h
@@ -300,7 +300,7 @@ struct Site  *krm_getFtypeSite(void);
 
 void    krm_releaseAllFtypeSites(struct Site *first_site_ptr);
  void    krm_releaseSiteArrays(void);
- void    krm_relocateLinkPtrs(int offset);
+ void    krm_relocateLinkPtrs(long int offset);
 	void	krm_releaseUnitArrays(void);
  int     krm_allocNTableArray(void);
 struct NameTable  *krm_getNTableEntry(void);
